+++ packages/js/src/index.umd.ts
@@ -1,4 +1,4 @@
-export * from './index';
+export * from './index_main';
 
 // Expose a stable namespace for UMD builds (window.lipilekhika.typing.*)
 export * as typing from './typing';